+++ lib/yt/models/asset.rb
@@ -12,11 +12,12 @@ module Yt
         @data = options.fetch(:data, {})
         @id = options[:id]
         @auth = options[:auth]
+        @params = options[:params]
       end
 
       def update(attributes = {})
         underscore_keys! attributes
-        do_patch body: attributes
+        do_update body: attributes
         true
       end
 
@@ -32,6 +33,10 @@ module Yt
 
       def metadata_effective
         @metadata_effective ||= Yt::Models::AssetMetadata.new data: @data.fetch('metadataEffective', {})
+      end
+
+      def ownership_effective
+        @ownership_effective ||= Yt::Models::Ownership.new data: @data.fetch('ownershipEffective', {})
       end
 
       # Soft-deletes the asset.
@@ -64,11 +69,11 @@ module Yt
       #   +'show'+, +'sound_recording'+, +'video_game'+, +'web'+.
       has_attribute :type
 
-      # @return [Array<Yt::Models::Tag>] the list of asset labels associated
+      # @return [Array<String>] the list of asset labels associated
       #   with the asset. You can apply a label to multiple assets to group
       #   them. You can use the labels as search filters to perform bulk updates,
       #   to download reports, or to filter YouTube Analytics.
-      has_attribute :label
+      has_attribute :label, default: []
 
 # Status
 
@@ -104,6 +109,24 @@ module Yt
           params[:expected_response] = Net::HTTPOK
           params[:path] = "/youtube/partner/v1/assets/#{@id}"
           params[:params] = {on_behalf_of_content_owner: @auth.owner_name}
+        end
+      end
+
+      # @see https://developers.google.com/youtube/partner/docs/v1/assets/update
+      def update_params
+        super.tap do |params|
+          params[:expected_response] = Net::HTTPOK
+          params[:path] = "/youtube/partner/v1/assets/#{@id}"
+          params[:params] = {on_behalf_of_content_owner: @auth.owner_name}
+        end
+      end
+
+      # @return [Hash] the parameters to submit to YouTube to get an asset.
+      # @see https://developers.google.com/youtube/partner/docs/v1/assets/get
+      def get_params
+        super.tap do |params|
+          params[:path] = "/youtube/partner/v1/assets/#{@id}"
+          params[:params] = {on_behalf_of_content_owner: @auth.owner_name}.merge! @params
         end
       end
     end