+++ lib/yt/models/playlist.rb
@@ -197,7 +197,8 @@ module Yt
       # @private
       # Tells `has_reports` to retrieve the reports from YouTube Analytics API
       # either as a Channel or as a Content Owner.
-      # @see https://developers.google.com/youtube/analytics/v1/reports
+      # @see https://developers.google.com/youtube/analytics/channel_reports
+      # @see https://developers.google.com/youtube/analytics/content_owner_reports
       def reports_params
         {}.tap do |params|
           if auth.owner_name
@@ -230,6 +231,15 @@ module Yt
         params.dup.tap do |params|
           params[:resource_id] ||= {kind: 'youtube#video', videoId: video_id}
         end
+      end
+
+      # For updating playlist with content owner auth.
+      # @see https://developers.google.com/youtube/v3/docs/playlists/update
+      def update_params
+        params = super
+        params[:params] ||= {}
+        params[:params].merge! auth.update_playlist_params
+        params
       end
     end
   end