+++ spec/models/request_spec.rb
@@ -7,6 +7,10 @@ describe Yt::Request do
   let(:response) { response_class.new nil, nil, nil }
   let(:response_body) { }
 
+  before do
+    allow(request).to receive(:retry_time).and_return(0)
+  end
+
   describe '#run' do
     context 'given a request that returns' do
       before { allow(response).to receive(:body).and_return response_body }
@@ -58,6 +62,15 @@ describe Yt::Request do
       end
 
       context 'an error code 401' do
+        let(:response_class) { Net::HTTPUnauthorized }
+
+        it { expect{request.run}.to fail }
+      end
+
+      context 'an error code 401 with a refresh token' do
+        before { expect(Net::HTTP).to receive(:start).twice.and_return response }
+        let(:auth) { double(refreshed_access_token?: true, access_token: 'whatever') }
+        let(:request) { Yt::Request.new host: 'example.com', auth: auth }
         let(:response_class) { Net::HTTPUnauthorized }
 
         it { expect{request.run}.to fail }