+++ spec/requests/as_server_app/comment_thread_spec.rb
@@ -2,15 +2,15 @@ require 'spec_helper'
 require 'yt/models/comment_thread'
 require 'yt/models/comment'
 
-describe Yt::CommentThread, :server_app do
+describe Yt::CommentThread, :server_app, :vcr do
   subject(:comment_thread) { Yt::CommentThread.new attrs }
 
   context 'given an existing comment thread ID about a channel' do
-    let(:attrs) { {id: 'z13kdnf4pursxnwr404cc3oz4zb0hjwirkg0k'} }
+    let(:attrs) { {id: 'UgzzJVW75s5KrSaf0Ah4AaABAg'} }
 
     it { expect(comment_thread.video_id).to be_nil }
     it { expect(comment_thread.total_reply_count).to be_an Integer }
-    it { expect(comment_thread.can_reply?).to be false }
+    it { expect(comment_thread.can_reply?).to be true }
     it { expect(comment_thread).to be_public }
 
     it { expect(comment_thread.top_level_comment).to be_a Yt::Comment }