+++ spec/requests/as_server_app/playlist_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 require 'yt/models/playlist'
 
-describe Yt::Playlist, :server_app do
+describe Yt::Playlist, :server_app, :vcr do
   subject(:playlist) { Yt::Playlist.new attrs }
 
   context 'given an existing playlist ID' do
-    let(:attrs) { {id: 'PLSWYkYzOrPMT9pJG5St5G0WDalhRzGkU4'} }
+    let(:attrs) { {id: 'PLpjK416fmKwQ6Ene4GKRLICznoE1QFyt4'} }
 
     it 'returns valid snippet data' do
       expect(playlist.snippet).to be_a Yt::Snippet
@@ -22,12 +22,5 @@ describe Yt::Playlist, :server_app do
     it { expect(playlist.status).to be_a Yt::Status }
     it { expect(playlist.playlist_items).to be_a Yt::Collections::PlaylistItems }
     it { expect(playlist.playlist_items.first).to be_a Yt::PlaylistItem }
-  end
-
-  context 'given an unknown playlist' do
-    let(:attrs) { {id: 'not-a-playlist-id'} }
-
-    it { expect{playlist.snippet}.to raise_error Yt::Errors::NoItems }
-    it { expect{playlist.status}.to raise_error Yt::Errors::NoItems }
   end
 end